Adding combined rain gauge data for years 2006-2007.
In the process I discovered two dates, one in each year, that had invalid formatting. They were both missing a slash '/' in the date dimension, and were discovered by formatting the cells in ISO format inside LibreOffice. These files have that year corrected.

This file also has the combined range dates in the 2006 year file averaged out so each entry corresponds to only one day. The last entry in 2006 on all gauges is combined for an 8 day range, this is rainfall measurement, so 8 days is a lot and it may be that most of the days were dry and one was extraordinarily rainy, but it could equally be that it was just as rainy one day as any of the others. Being from Mississippi myself I'm going to bet it was basically the same heavy rain for most of the time and spread it equally over the days.